Refactor the code to match MVC-structure
