+++ README.md
@@ -2,5 +2,6 @@
 
 Länk till repository: https://github.com/SamanMau/Tetris_projekt_G19
 
-Instruktioner: Extrahera mappen Tetris_projekt_G19.zip. Öppna en java idea (intellij). Öppna mappen Tetris. Öppna main.java i src mappen och kör programmet.
+Instruktioner: Extrahera mappen Tetris_projekt_G19.zip. Öppna en java idea (intellij). Öppna mappen Tetris. Öppna main.java i src mappen och kör programmet. 
+När programmet startas ser man en block som befinner sig längst upp på spelplanen. För att blocken ska flytta ner klick på "Start Game".
