+++ Tetris/src/Main.java
@@ -1,7 +1,7 @@
-import view.*;
+import Control.Controller;
+import View.*;
 public class Main {
     public static void main(String[] args) {
-        MainFrame frame = new MainFrame();
-
+        Controller controller = new Controller();
     }
 }
