+++ Tetris/src/Model/ListOfBlocks.java
@@ -1,7 +1,16 @@
+/**
+ * A class used to manage a list of blocks.
+ * Den har en instansvariabel för varje
+ */
+
 package Model;
 import java.awt.*;
 import java.util.*;
-
+/**
+ * This class manage a list of blocks for a game.
+ * Each block is represented by a 2D array of integers where 1 represents a block cell and 0 represents an empty cell.
+ * This class also manages a list of colors related to each block.
+ */
 public class ListOfBlocks  {
     private int[][] tBlock;
     private int[][] lBlock;
@@ -13,7 +22,10 @@ public class ListOfBlocks  {
     private ArrayList<int[][]> listOfBlocks;
     private ArrayList<Color> listOfColors;
 
-
+    /**
+     * Create blocks and add them to the list.
+     * Also create the list of colors.
+     */
     public ListOfBlocks(){
         createBlock();
         listOfBlocks = new ArrayList<>();
@@ -32,7 +44,9 @@ public class ListOfBlocks  {
         oBlock = new int[][]{{1,1}, {1,1}};
 
     }
-
+    /**
+     * Initializes all block.
+     */
     public void addBlockToList(){
         listOfBlocks.add(tBlock);
         listOfBlocks.add(lBlock);
@@ -42,7 +56,9 @@ public class ListOfBlocks  {
         listOfBlocks.add(oBlock);
         listOfBlocks.add(jBlock);
     }
-
+    /**
+     * Add to the list of colors.
+     */
     public void addColorsToList(){
         listOfColors.add(Color.white);
         listOfColors.add(Color.orange);
@@ -52,8 +68,20 @@ public class ListOfBlocks  {
         listOfColors.add(Color.pink);
         listOfColors.add(Color.yellow);
     }
+
+    /**
+     * Gets the list of block shape.
+     *
+     * @return The list of block.
+     */
     public ArrayList<int[][]> getBlockList(){
         return listOfBlocks;
     }
+
+    /**
+     * Gets the list of colors.
+     *
+     * @return The list of colors.
+     */
     public ArrayList<Color> getListOfColors(){return listOfColors;}
 }
