+++ Tetris/src/Model/TetrisBlock.java
@@ -1,11 +1,22 @@
+/**
+ * A model class that represents a tetris block.
+ * The class has relevant getters to manage changes in coordinates in the gameArea.
+ * Study the methods and see in which context they are used (find usages).
+ */
 package Model;
 import java.awt.*;
 
 public class TetrisBlock  {
     private int[][] shape;
     private Color color;
-    private int x;
-    private int y;
+    private int x; // is used to determine where in the game plan the tetris block is going to move in the x coordinate.
+    private int y; // is used to determine where in the game plan the tetris block is going to move in the y coordinate.
+
+    /**
+     * Constructor used to create objects of a tetris block.
+     * @param shape when creating a block, you enter 1:s and 0:s in a 2d array.
+     * @param color and then you give it a color.
+     */
    public TetrisBlock(int[][] shape, Color color){
        createBlock(shape, color);
        this.x = 4;
@@ -24,6 +35,11 @@ public class TetrisBlock  {
    public Color getColor(){
        return this.color;
    }
+
+    /**
+     * The methods below are used to increment or decrement x and y coordinates so that
+     * we are able to move a block in tehe gameArea.
+     */
 
     public void incrementY() {
         y++;
