+++ Tetris/src/View/BottomPanel.java
@@ -1,11 +1,11 @@
-package view;
+/**
+ * A class that will be used to add an image of our choice to the GUI.
+ */
+package View;
 
-import javax.imageio.ImageIO;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.image.BufferedImage;
-import java.io.File;
-import java.io.IOException;
 
 public class BottomPanel extends JPanel {
     private BufferedImage testImage;
