+++ Tetris/src/View/Buttons.java
@@ -1,4 +1,4 @@
-package view;
+package View;
 
 public enum Buttons {
     startaSpel, avslutaSpel, visaHighScore;
