+++ Tetris/src/View/LPanel.java
@@ -1,7 +1,9 @@
-package view;
+/**
+ * A class that represents a left panel.
+ */
+package View;
 
 import javax.swing.*;
-import javax.swing.border.*;
 import java.awt.*;
 
 public class LPanel extends JPanel {
@@ -15,11 +17,22 @@ public class LPanel extends JPanel {
         this.setVisible(true);
     }
 
+    /**
+     * Method used to make a text field that represents the users points,
+     * this will increase everytime the user gets more points.
+     * @param point when a users points changes, we use this parameter to add it to the text field.
+     */
     public void createPointCounter(int point){
-        poangText.setText("Poäng: " + point);
+        poangText.setText("Points: " + point);
         poangText.setFont(new Font("Calibri", Font.BOLD, 20));
         this.add(poangText);
     }
+
+    /**
+     * Method used to make a text field that represents the users level,
+     * this will increase everytime the user reaches a new level.
+     * @param level when a users level changes, we use this parameter to add it to the text field.
+     */
     public void createLevelCounter(int level){
         levelText.setText("Level: " + level);
         levelText.setFont(new Font("Calibri", Font.BOLD, 20));
